Replace LinkUserTemplate in api module  (#112)
* Replace LinkUserTemplate#of with LinkUser#of

* Remove LinkUserTemplate

* Remove unnecessary field declaration in AccountLinkAdapter
