+++ core/src/main/java/me/mastercapexd/auth/link/user/LinkUserTemplate.java
@@ -5,13 +5,15 @@ import com.bivashy.auth.api.link.LinkType;
 import com.bivashy.auth.api.link.user.LinkUser;
 import com.bivashy.auth.api.link.user.info.LinkUserInfo;
 
-public abstract class LinkUserTemplate implements LinkUser {
-    protected final LinkType linkType;
-    protected final Account account;
+public class LinkUserTemplate implements LinkUser {
+    private final LinkType linkType;
+    private final Account account;
+    private final LinkUserInfo linkUserInfo;
 
-    public LinkUserTemplate(LinkType linkType, Account account) {
+    public LinkUserTemplate(LinkType linkType, Account account, LinkUserInfo linkUserInfo) {
         this.linkType = linkType;
         this.account = account;
+        this.linkUserInfo = linkUserInfo;
     }
 
     @Override
@@ -24,12 +26,12 @@ public abstract class LinkUserTemplate implements LinkUser {
         return account;
     }
 
+    @Override
+    public LinkUserInfo getLinkUserInfo() {
+        return linkUserInfo;
+    }
+
     public static LinkUser of(LinkType linkType, Account account, LinkUserInfo linkUserInfo) {
-        return new LinkUserTemplate(linkType, account) {
-            @Override
-            public LinkUserInfo getLinkUserInfo() {
-                return linkUserInfo;
-            }
-        };
+        return new LinkUserTemplate(linkType, account, linkUserInfo);
     }
 }
